requirements.txt
antiorm==1.2.1
appdirs==1.4.4
asgiref==3.3.4
click==7.1.2
colorama==0.4.4
db==0.1.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-heroku==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.24
mysqlclient==2.0.3
Pillow==8.2.0
pilo==0.6.2
psycopg2==2.8.6
pyodbc==4.0.30
python-decouple==3.4
pytz==2021.1
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.4
Werkzeug==1.0.1
whitenoise==5.2.0
